Addressed failing checks
